fixes after labs - part 1.

+++ accstruct.h
@@ -4,9 +4,9 @@
 #define STRING_LIMIT 20
 
 typedef struct {
-	unsigned long accountNumber;
-	unsigned long pesel;
-	unsigned long balance;
+	unsigned long long accountNumber;
+	unsigned long long pesel;
+	unsigned long long balance;
 	char name[STRING_LIMIT + 1];
 	char lastName[STRING_LIMIT + 1];
 	char address[2 * STRING_LIMIT + 1];
